upload codebase

+++ run_eval.py
@@ -99,8 +99,8 @@ class Trainer:
                 # inputialize the output map
                 h, w = testdataloader.dataset.shape()
                 output_map = torch.zeros((h, w), dtype=torch.float32)
-                output_scale_map = torch.zeros((h, w), dtype=torch.float16)
-                output_map_count = torch.zeros((h, w), dtype=torch.int8)
+                output_scale_map = torch.zeros((h, w), dtype=torch.float32)
+                output_map_count = torch.zeros((h, w), dtype=torch.int16)
 
                 # if len(self.model) > 1:
                 output_map_squared = torch.zeros((h, w), dtype=torch.float32)
@@ -119,7 +119,7 @@ class Trainer:
 
                     # get the output with a forward pass
                     popdense = torch.zeros((len(self.model), ips, ips), dtype=torch.float32, device="cuda")
-                    scale = torch.zeros((len(self.model), ips, ips), dtype=torch.float16, device="cuda")
+                    scale = torch.zeros((len(self.model), ips, ips), dtype=torch.float32, device="cuda")
                     popdense_squared = torch.zeros((len(self.model), ips, ips), dtype=torch.float32, device="cuda")
                     scale_squared = torch.zeros((len(self.model), ips, ips), dtype=torch.float32, device="cuda")
                     for i, model in enumerate(self.model):
@@ -146,7 +146,7 @@ class Trainer:
 
                     if "scale" in output.keys():
                         if output["scale"] is not None:
-                            output_scale_map[xl:xl+ips, yl:yl+ips][mask.cpu()] += output["scale"][0][mask].cpu().to(torch.float16)
+                            output_scale_map[xl:xl+ips, yl:yl+ips][mask.cpu()] += output["scale"][0][mask].cpu().to(torch.float32)
                             output_scale_map_squared[xl:xl+ips, yl:yl+ips][mask.cpu()] += output["scale_squared"][0][mask].cpu().to(torch.float32)
 
                     output_map_count[xl:xl+ips, yl:yl+ips][mask.cpu()] += len(self.model)
@@ -157,7 +157,7 @@ class Trainer:
                 div_mask = output_map_count > 1
                 # a = output_map.clone()
 
-                output_map[div_mask] = output_map[div_mask] / output_map_count[div_mask]
+                output_map[div_mask] = output_map[div_mask] / output_map_count[div_mask].to(torch.float32)
                 # eps = 1e-10
 
                 # calculate the standard deviation from the sum of squares and the mean as "std_dev = math.sqrt((sum_of_squares - n * mean ** 2) / (n - 1))"
